feat: extract the ai generation button and logic out of the BlockIconHeader
