+++ frontend/src/lib/components/blocks/BlockIconHeader.svelte
@@ -7,13 +7,8 @@
 		THEORY_BLOCK_TYPE
 	} from '$lib/schemas/response/BlockRes';
 	import BlockIcon from './BlockIcon.svelte';
-	import { Sparkles } from 'lucide-svelte';
-	import CreateMultipleChoiceModal from '../dialogs/CreateMultipleChoiceModal.svelte';
-	import { api } from '$lib/api/apiClient';
-	import { getLearningUnitById } from '$lib/api/collections/learningUnit';
-	import { createQuery } from '@tanstack/svelte-query';
 	import { INSTRUCTOR_ROLE, type RoleType } from '$lib/schemas/response/UserInfo';
-	import GenerateTheoryModal from '../GenerateTheoryModal.svelte';
+	import BlockAiGenerationButton from '$lib/components/blocks/BlockAiGenerationButton.svelte';
 
 	interface BlockIconHeaderProps {
 		block: BlockRes;
@@ -21,18 +16,6 @@
 		role: RoleType;
 	}
 	const { block, learningUnitId: learningUnitId, role }: BlockIconHeaderProps = $props();
-
-	const getLearningKit = createQuery({
-		queryKey: ['learning-unit', learningUnitId],
-		enabled: !!learningUnitId,
-		queryFn: () => {
-			if (learningUnitId) {
-				return api(fetch).req(getLearningUnitById, null, learningUnitId).parse();
-			}
-		}
-	});
-
-	let theoryBlocks: BlockRes[] = $state([]);
 
 	let blockTypeTerm = $derived.by(() => {
 		switch (block.type) {
@@ -46,23 +29,9 @@
 				return 'Unknown Block';
 		}
 	});
-
-	let showCreationDialog = $state(false);
-
-	function handleCreationDialog() {
-		showCreationDialog = false;
-	}
-
-	$effect(() => {
-		if ($getLearningKit.isSuccess && $getLearningKit?.data) {
-			theoryBlocks = $getLearningKit?.data.blocks.filter(
-				(b: BlockRes) => b.type === THEORY_BLOCK_TYPE
-			);
-		}
-	});
 </script>
 
-<div class="flex w-full items-center justify-between">
+<div class="flex justify-between">
 	<div class="flex items-center gap-2">
 		<BlockIcon iconType={block.type} />
 		<div class="flex items-baseline gap-2">
@@ -72,41 +41,8 @@
 	</div>
 
 	{#if learningUnitId && role === INSTRUCTOR_ROLE}
-		<button
-			type="button"
-			class="btn preset-filled-primary-500"
-			title={$_('block.generateAi')}
-			onclick={(e) => {
-				e.preventDefault();
-				showCreationDialog = true;
-			}}
-		>
-			{$_('block.generateAi')}
-			<Sparkles size={18} />
-		</button>
+		<div class="ml-auto">
+			<BlockAiGenerationButton {block} {learningUnitId} />
+		</div>
 	{/if}
 </div>
-
-{#if block.type === 'MULTIPLE_CHOICE'}
-	<CreateMultipleChoiceModal
-		isOpen={showCreationDialog}
-		onConfirm={handleCreationDialog}
-		onCancel={() => (showCreationDialog = false)}
-		theoryBlocks={theoryBlocks.map((theoryBlock) => ({
-			id: theoryBlock.uuid,
-			title: theoryBlock.name
-		}))}
-	/>
-{:else if block.type === 'THEORY'}
-	<GenerateTheoryModal bind:isOpen={showCreationDialog} blockId={block.uuid} />
-{:else if block.type === 'QUESTION'}
-	<CreateMultipleChoiceModal
-		isOpen={showCreationDialog}
-		onConfirm={handleCreationDialog}
-		onCancel={() => (showCreationDialog = false)}
-		theoryBlocks={theoryBlocks.map((theoryBlock) => ({
-			id: theoryBlock.uuid,
-			title: theoryBlock.name
-		}))}
-	/>
-{/if}
